chore: keep delete modal as component
